requirements.txt
albumentations==1.3.1
fastapi==0.101.0
lanyocr==0.1.12
moviepy==1.0.3
numpy==1.24.3
opencv_python==4.8.0.74
pandas==2.0.1
passlib==1.7.4
Pillow==9.4.0
pydantic==1.10.7
python_jose==3.3.0
Requests==2.31.0
SQLAlchemy==2.0.15
streamlit==1.22.0
supervision==0.6.0
torch==2.0.0
ultralytics==8.0.145
